backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "ResumAI-TPCC",email = "resumaitpcc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (==0.128.0)",
    "uvicorn[standard] (==0.27.0)",
    "httpx (==0.26.0)",
    "google-cloud-storage (==2.16.0)",
    "python-dotenv (==1.0.0)",
    "pydantic-settings (==2.1.0)",
    "python-multipart (==0.0.18)",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "google-generativeai (>=0.8.0)",
    "pdfplumber (>=0.11.0)",
    "python-docx (>=1.1.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"